+++ src/utils/index.ts
@@ -1,4 +1,3 @@
 export { default as errorObjToString } from './errorObjToString';
 export { default as extractUsernameFromEmail } from './extractUsernameFromEmail';
-export { default as reqTryCatch } from './reqTryCatch';
 export { default as requiredFormValue } from './requiredFormValue';
